requirements-dev.txt
opencv-python
python-Levenshtein
pytesseract
